check for correct owner before executing

+++ Comments/removeComment.php
@@ -2,13 +2,20 @@
     require("../connect.php");
     require("../requireLogin.php");
 
-    $sql = $conn->prepare("DELETE FROM comments WHERE user_id=? AND game_id=? AND addedTime=?");
-    try{
-        $sql->execute(array($_SESSION['user_id'], $_POST['game_id'], $_POST['time']));
+    // owner verification
+    $sql_verify = $conn->prepare("SELECT user_id FROM comments WHERE game_id=?");
+    $sql_verify->execute(array($_POST['game_id']));
+    $result_verfy = $sql_verify->fetch();
+    
+    if($_SESSION['user_id'] != $result_verfy['user_id']){
         header("Location: /game-library/game/gameDetails.php?game_id=".$_POST['game_id']);
-    }catch(PDOException $err){
-        echo "Error";
+    }else{
+        try{
+            $sql = $conn->prepare("DELETE FROM comments WHERE user_id=? AND game_id=? AND addedTime=?");
+            $sql->execute(array($_SESSION['user_id'], $_POST['game_id'], $_POST['time']));
+            header("Location: /game-library/game/gameDetails.php?game_id=".$_POST['game_id']);
+        }catch(PDOException $err){
+            echo "Error";
+        }
     }
-
-
 ?>
